refactor: 패키지 구조를 세분화한다.
