+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.Di"
         tools:targetApi="31">
         <activity
-            android:name=".injectionexample.fragment.ExampleActivity"
+            android:name=".injectionexample.fragment.sharedViewModel.ExampleActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
