--- app/src/main/java/com/boogiwoogi/di/injectionexample/fragment/ExampleRepository.kt
@@ -1,6 +1,0 @@
-package com.boogiwoogi.di.injectionexample.fragment
-
-interface ExampleRepository {
-
-    fun fetchData(): List<Int>
-}
