+++ app/src/main/java/com/boogiwoogi/di/injectionexample/fragment/sharedViewModel/DefaultExampleRepository.kt
@@ -1,4 +1,4 @@
-package com.boogiwoogi.di.injectionexample.fragment
+package com.boogiwoogi.di.injectionexample.fragment.sharedViewModel
 
 class DefaultExampleRepository : ExampleRepository {
