+++ app/src/main/java/com/boogiwoogi/di/injectionexample/fragment/sharedViewModel/ExampleRepository.kt
@@ -1,0 +1,6 @@
+package com.boogiwoogi.di.injectionexample.fragment.sharedViewModel
+
+interface ExampleRepository {
+
+    fun fetchData(): List<Int>
+}
